ray-gen.py
import os
import argparse
import torch
from diffusers import StableDiffusionPipeline
from torchvision import transforms, datasets, models
from torch.utils.data import DataLoader
import torch.nn as nn
import torch.optim as optim
import pandas as pd
from sklearn.metrics import roc_auc_score
from PIL import Image

