requirements.txt
torch>=1.13.0
torchvision>=0.14.0
diffusers[torch]>=0.11.0
transformers>=4.26.0
scikit-learn>=1.0.0
pandas>=1.4.0
pillow>=9.0.0
accelerate>=0.10.0
pydicom==2.4.4
huggingface_hub
peft 
bitsandbytes